+++ src/server/index.js
@@ -8,32 +8,18 @@ const HttpRouter = require("./router/http");
 const app = Websockify(new Koa());
 const rootPath = path.join(__dirname, "../../frontend/");
 const { setup } = require("../link/setup");
-
 const { logger } = require("../util");
 
-/*
- ===================================
- WebSocket Router
- ===================================
- */
-exports.start = function(port, cb) {
+exports.start = (port, cb) => {
   setup();
   app.use(bodyParser());
-  app.ws.use(WsRouter.routes());
-  app.on("error", function(err, ctx) {
-    if (err.status === 404) {
-      logger.verbose(err);
-    } else {
-      logger.verbose(err);
-    }
-  });
-  /*
-    ===================================
-    Http Router
-    ===================================
-  */
+  app.ws.use(WsRouter.routes()).use(WsRouter.allowedMethods());
   app.use(HttpRouter.routes());
   app.use(serve(rootPath));
 
+  app.on("error", (err, ctx) => {
+    logger.verbose(err);
+  });
+  
   app.listen(port, cb);
 };